Provide an opt-out for disabling the optimization of generics serialization.
The existing optimization of generics serialization aims at reducing the size of the serialized representation. This optimization is enabled by default. But sometimes it may have a negative impact on performance. And sometimes it cannot handle complex cases of generics usage.

This patch provides a mechanism to opt-out from using this optimization. To disable it, simply invoke `FieldSerializerConfig.setOptimizedGenerics(false)`.